sound: Nuke CONFIG_SND_SUNXI_SOC_SUPPORT_AUDIO_RAW option
If this option was set, specialy modified alsa-lib was needed to be able to use ALSA. Raw audio interface can still be set by mixer settings, except for audiohub. Not much is lost, except drop-in compatibility with Android, which was newer our intention. Later, mixer settings can be replaced with standard IEC958 interface, so raw audio would be completely compatible with many Linux players.

Signed-off-by: Jernej Skrabec <jernej.skrabec@siol.net>